Fix c++ configurations. Set max m1/m2 current to prevent latching error.

+++ TeensyV2/modules/roboclaw/RoboClaw.cpp
@@ -532,7 +532,7 @@ bool RoboClaw::ResetEncoders(uint8_t address) {
 }
 
 bool RoboClaw::ReadVersion(uint8_t address, char *version) {
-  uint8_t data;
+  int data;
   uint8_t trys = MAXRETRY;
   do {
     flush();
